bytecler/requirements.txt
telethon>=1.36.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
